requirements.txt
gunicorn
flask
flask-wtf
flask-bcrypt
Pillow
Markdown
requests-oauthlib==1.3.0
Janome==0.3.10
wordcloud==1.7.0
tensorflow-gpu==1.15
googledrivedownloader
tqdm